修改11

+++ app/src/main/java/com/hippo/ehviewer/ui/scene/DownloadsScene.java
@@ -552,6 +552,10 @@ public class DownloadsScene extends ToolbarScene
                 }
                 return true;
             }
+            case R.id.action_open_download_labels: {
+                openDrawer(Gravity.RIGHT);
+                return true;
+            }
             case R.id.action_reset_reading_progress: {
                 new MaterialAlertDialogBuilder(getContext())
                         .setMessage(R.string.reset_reading_progress_message)
@@ -1058,7 +1062,7 @@ public class DownloadsScene extends ToolbarScene
                 });
                 if (position > 0) {
                     holder.option.setVisibility(View.VISIBLE);
-                    holder.option.setOnClickListener(v -> {
+                    holder.itemView.setOnLongClickListener(v -> {
                         if (context != null) {
                             PopupMenu popupMenu = new PopupMenu(context, holder.option);
                             popupMenu.inflate(R.menu.download_label_option);
@@ -1077,14 +1081,22 @@ public class DownloadsScene extends ToolbarScene
                                         }
                                         break;
                                     case R.id.menu_label_remove:
-                                        mDownloadManager.deleteLabel(label);
-                                        mLabels.remove(position);
-                                        notifyDataSetChanged();
+                                        new MaterialAlertDialogBuilder(requireContext())
+                                                .setTitle(getString(R.string.delete_label_title))
+                                                .setMessage(getString(R.string.delete_label_message, label))
+                                                .setPositiveButton(R.string.delete, (dialog, which) -> {
+                                                    mDownloadManager.deleteLabel(label);
+                                                    mLabels.remove(position);
+                                                    notifyDataSetChanged();
+                                                })
+                                                .setNegativeButton(android.R.string.cancel, null)
+                                                .show();
                                         break;
                                 }
                                 return false;
                             });
                         }
+                        return true;
                     });
                 } else {
                     holder.option.setVisibility(View.GONE);
@@ -1104,7 +1116,7 @@ public class DownloadsScene extends ToolbarScene
 
         @Override
         public boolean onCheckCanStartDrag(@NonNull DownloadLabelHolder holder, int position, int x, int y) {
-            return position != 0;
+            return position != 0 && x > holder.option.getX() && y > holder.option.getY();
         }
 
         @Override
@@ -1576,8 +1588,9 @@ public class DownloadsScene extends ToolbarScene
                 mBuilder.setError(null);
                 mDialog.dismiss();
                 EhApplication.getDownloadManager(context).addLabel(text);
-                if (mAdapter != null && mLabels != null) {
-                    mAdapter.notifyItemInserted(mLabels.size() - 1);
+                initLabels();
+                if (mLabelAdapter != null && mLabels != null) {
+                    mLabelAdapter.notifyItemInserted(mLabels.size() - 1);
                 }
                 if (mViewTransition != null) {
                     if (mLabels != null && mLabels.size() > 0) {
@@ -1586,10 +1599,6 @@ public class DownloadsScene extends ToolbarScene
                         mViewTransition.showView(1);
                     }
                 }
-            }
-            if (mLabelAdapter != null) {
-                initLabels();
-                mLabelAdapter.notifyDataSetChanged();
             }
         }
     }
